backend/requirements.txt
appdirs==1.4.4
asgiref==3.7.2
black==23.9.1
blessed==1.20.0
bpython==0.24
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
curtsies==0.4.2
cwcwidth==0.1.8
Django==4.2.4
django-extensions==3.2.3
django-filter==23.2
djangorestframework==3.14.0
greenlet==2.0.2
idna==3.4
install==1.3.5
jedi==0.19.0
Markdown==3.4.4
mypy-extensions==1.0.0
packaging==23.1
parso==0.8.3
pathspec==0.11.2
platformdirs==3.10.0
prompt-toolkit==3.0.39
ptpython==3.0.23
Pygments==2.16.1
pytz==2023.3.post1
pyxdg==0.28
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.4
wcwidth==0.2.6
